backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
ollama>=0.3.0
chromadb>=0.4.22
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
redis==5.0.1
celery==5.3.6
pytest==7.4.4
